Fix all tests
